fix(davinci-client): break circular import by extracting collector resolver
