--- goals/centralize-redux-store/facts-review.json
@@ -1,103 +1,0 @@
-{
-  "stage": "facts",
-  "title": "Centralize Redux Store Across Client Packages",
-  "goalSlug": "centralize-redux-store",
-  "facts": [
-    {
-      "id": "fact-1",
-      "text": "`davinci()` returns an object with a `store` property typed as an opaque `SdkStore` handle (not the raw Redux `Store` type).",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-2",
-      "text": "`journey()` returns an object with a `store` property typed as an opaque `SdkStore` handle.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-3",
-      "text": "`oidc()` accepts an optional second argument — an `SdkStore` — and uses it instead of creating a new store when provided.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-4",
-      "text": "When `oidc()` is called without a second argument, it creates its own store exactly as it does today (no behavioral change for standalone usage).",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-5",
-      "text": "The davinci/journey store is pre-provisioned with oidc's reducers (`oidcApi`) so that when `oidc()` receives the shared store, it can dispatch and select from its own slice without any runtime injection.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-6",
-      "text": "`wellknownApi` is extracted from `davinci-client`, `journey-client`, and `oidc-client` into `sdk-effects` (or an appropriate shared effects package) so all three import the same instance.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-7",
-      "text": "When `oidc()` shares a store with `davinci()`, there is only one `wellknown` cache entry — a second call to the same wellknown URL returns from cache without a network request.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-8",
-      "text": "A single `store.subscribe()` call on the davinci/journey client fires for state changes from both the flow client and oidc client when a shared store is in use.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-9",
-      "text": "All existing call sites (`davinci({ config })`, `journey({ config })`, `oidc({ config })`) continue to work without modification.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-10",
-      "text": "TypeScript consumers cannot access internal Redux primitives (dispatch, getState, etc.) through the exposed `SdkStore` type — it is opaque by design.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    },
-    {
-      "id": "fact-11",
-      "text": "The `requestMiddleware` and `logger` passed to `davinci()`/`journey()` are inherited by oidc when using a shared store — oidc does not need to re-declare them.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": false,
-      "automatedVerification": false
-    },
-    {
-      "id": "fact-12",
-      "text": "All affected packages pass `pnpm lint`, `pnpm build`, and `pnpm test` after the changes.",
-      "accepted": false,
-      "removed": false,
-      "recommendedAutomatedVerification": true,
-      "automatedVerification": true
-    }
-  ]
-}
