--- goals/centralize-redux-store/facts.md
@@ -1,14 +1,0 @@
-# Facts
-
-- `davinci()` returns an object with a `store` property typed as an opaque `SdkStore` handle (not the raw Redux `Store` type).
-- `journey()` returns an object with a `store` property typed as an opaque `SdkStore` handle.
-- `oidc()` accepts an optional second argument — an `SdkStore` — and uses it instead of creating a new store when provided.
-- When `oidc()` is called without a second argument, it creates its own store exactly as it does today (no behavioral change for standalone usage).
-- `oidc` lazily injects its own reducers (`oidcApi`) into the shared store at `oidc()` init time — the owning store (davinci/journey) does NOT pre-provision oidc reducers.
-- `wellknownApi` is extracted from `davinci-client`, `journey-client`, and `oidc-client` into `sdk-effects` (or an appropriate shared effects package) so all three import the same instance.
-- When `oidc()` shares a store with `davinci()`, there is only one `wellknown` cache entry — a second call to the same wellknown URL returns from cache without a network request.
-- A single `store.subscribe()` call on the davinci/journey client fires for state changes from both the flow client and oidc client when a shared store is in use.
-- All existing call sites (`davinci({ config })`, `journey({ config })`, `oidc({ config })`) continue to work without modification.
-- TypeScript consumers cannot access internal Redux primitives (dispatch, getState, etc.) through the exposed `SdkStore` type — it is opaque by design.
-- The `requestMiddleware` and `logger` passed to `davinci()`/`journey()` are inherited by oidc when using a shared store — oidc does not need to re-declare them.
-- All affected packages pass `pnpm lint`, `pnpm build`, and `pnpm test` after the changes.
