--- goals/centralize-redux-store/facts.meta.json
@@ -1,86 +1,0 @@
-[
-  {
-    "id": "fact-1",
-    "text": "`davinci()` returns an object with a `store` property typed as an opaque `SdkStore` handle (not the raw Redux `Store` type).",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-2",
-    "text": "`journey()` returns an object with a `store` property typed as an opaque `SdkStore` handle.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-3",
-    "text": "`oidc()` accepts an optional second argument — an `SdkStore` — and uses it instead of creating a new store when provided.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-4",
-    "text": "When `oidc()` is called without a second argument, it creates its own store exactly as it does today (no behavioral change for standalone usage).",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-5-revised",
-    "text": "`oidc` lazily injects its own reducers (`oidcApi`) into the shared store at `oidc()` init time — the owning store (davinci/journey) does NOT pre-provision oidc reducers.",
-    "comment": "User rejected pre-provisioning: 'I don't think we want oidc pre-provisioned. We only want oidc when it's been used as the oidc client.'",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-6",
-    "text": "`wellknownApi` is extracted from `davinci-client`, `journey-client`, and `oidc-client` into `sdk-effects` (or an appropriate shared effects package) so all three import the same instance.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-7",
-    "text": "When `oidc()` shares a store with `davinci()`, there is only one `wellknown` cache entry — a second call to the same wellknown URL returns from cache without a network request.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-8",
-    "text": "A single `store.subscribe()` call on the davinci/journey client fires for state changes from both the flow client and oidc client when a shared store is in use.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-9",
-    "text": "All existing call sites (`davinci({ config })`, `journey({ config })`, `oidc({ config })`) continue to work without modification.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-10",
-    "text": "TypeScript consumers cannot access internal Redux primitives (dispatch, getState, etc.) through the exposed `SdkStore` type — it is opaque by design.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  },
-  {
-    "id": "fact-11",
-    "text": "The `requestMiddleware` and `logger` passed to `davinci()`/`journey()` are inherited by oidc when using a shared store — oidc does not need to re-declare them.",
-    "comment": "",
-    "recommendedAutomatedVerification": false,
-    "automatedVerification": false
-  },
-  {
-    "id": "fact-12",
-    "text": "All affected packages pass `pnpm lint`, `pnpm build`, and `pnpm test` after the changes.",
-    "comment": "",
-    "recommendedAutomatedVerification": true,
-    "automatedVerification": true
-  }
-]
