--- goals/centralize-redux-store/interview-result.json
@@ -1,53 +1,0 @@
-{
-  "decision": "submitted",
-  "stage": "interview",
-  "result": {
-    "stage": "interview",
-    "title": "Centralize Redux Store Across Client Packages",
-    "goalSlug": "centralize-redux-store",
-    "answers": [
-      {
-        "questionId": "store-exposure",
-        "selectedOptionIds": ["opaque-property"],
-        "customAnswer": "",
-        "answer": "store property (opaque SdkStore type)",
-        "completed": true
-      },
-      {
-        "questionId": "oidc-injection",
-        "selectedOptionIds": ["second-arg"],
-        "customAnswer": "",
-        "answer": "Second argument to oidc()",
-        "completed": true
-      },
-      {
-        "questionId": "wellknown-dedup",
-        "selectedOptionIds": ["extract-now"],
-        "customAnswer": "",
-        "answer": "Extract wellknownApi to sdk-effects in this PR",
-        "completed": true
-      },
-      {
-        "questionId": "reducer-injection-strategy",
-        "selectedOptionIds": ["pre-register"],
-        "customAnswer": "",
-        "answer": "Pre-register oidc reducers in davinci/journey store upfront",
-        "completed": true
-      },
-      {
-        "questionId": "scope",
-        "selectedOptionIds": ["davinci", "journey", "oidc", "sdk-effects"],
-        "customAnswer": "",
-        "answer": "davinci-client\njourney-client\noidc-client\nsdk-effects (wellknown extraction)",
-        "completed": true
-      },
-      {
-        "questionId": "breaking-changes",
-        "selectedOptionIds": ["minor-ok"],
-        "customAnswer": "",
-        "answer": "Minor breaking changes acceptable (semver minor)",
-        "completed": true
-      }
-    ]
-  }
-}
