--- goals/centralize-redux-store/interview.json
@@ -1,113 +1,0 @@
-{
-  "stage": "interview",
-  "title": "Centralize Redux Store Across Client Packages",
-  "goalSlug": "centralize-redux-store",
-  "questions": [
-    {
-      "id": "store-exposure",
-      "prompt": "How should `davinci()`/`journey()` expose the store to consumers?",
-      "description": "The store is currently fully private. The simplest approach is adding a `store` property to the returned client object. An alternative is a dedicated `getStore()` method. A third option is exposing only an opaque handle typed as `SdkStore` (not the full Redux `Store` type) to discourage direct Redux usage.",
-      "answerMode": "single-custom",
-      "recommendedAnswer": "Add a `store` property typed as an opaque `SdkStore` handle on the return object — additive, zero breaking changes, consumers can pass it to oidc() but can't easily misuse the raw Redux API.",
-      "recommendedOptionIds": ["opaque-property"],
-      "options": [
-        {
-          "id": "opaque-property",
-          "label": "store property (opaque SdkStore type)"
-        },
-        {
-          "id": "full-property",
-          "label": "store property (full Redux Store type)"
-        },
-        { "id": "getter-method", "label": "getStore() method" }
-      ],
-      "required": true
-    },
-    {
-      "id": "oidc-injection",
-      "prompt": "How should `oidc()` accept the shared store?",
-      "description": "Option A: add an optional `store` param to the existing config object — `oidc({ config, store: client.store })`. Option B: add it as a second argument — `oidc({ config }, client.store)`. The survey shows oidc() already takes a single destructured options bag.",
-      "answerMode": "single-custom",
-      "recommendedAnswer": "Optional `store` field inside the existing options object — consistent with how requestMiddleware and logger are passed today.",
-      "recommendedOptionIds": ["options-bag"],
-      "options": [
-        {
-          "id": "options-bag",
-          "label": "Optional field in existing options object"
-        },
-        { "id": "second-arg", "label": "Second argument to oidc()" }
-      ],
-      "required": true
-    },
-    {
-      "id": "wellknown-dedup",
-      "prompt": "Should `wellknownApi` be deduplicated as part of this work?",
-      "description": "All three packages each define their own `wellknownApi` RTK Query slice with the same reducerPath ('wellknown') and endpoint. With a shared store, both registrations would collide on the same reducer key. We could: (A) keep them separate and let RTK Query's cache dedup on-the-fly (they share the same reducerPath so this actually works), (B) extract `wellknownApi` to `sdk-effects` so both packages import the same instance, or (C) defer this to a follow-up.",
-      "answerMode": "single-custom",
-      "recommendedAnswer": "Extract wellknownApi to sdk-effects as part of this PR — the same-key collision is a latent bug we'd hit immediately when testing shared store, and extraction is straightforward.",
-      "recommendedOptionIds": ["extract-now"],
-      "options": [
-        {
-          "id": "extract-now",
-          "label": "Extract wellknownApi to sdk-effects in this PR"
-        },
-        {
-          "id": "rtk-dedup",
-          "label": "Rely on RTK Query same-key dedup (same reducerPath)"
-        },
-        { "id": "defer", "label": "Defer to a follow-up" }
-      ],
-      "required": true
-    },
-    {
-      "id": "reducer-injection-strategy",
-      "prompt": "How should oidc inject its reducers into a shared store?",
-      "description": "RTK 2.0 ships `combineSlices` and lazy reducer injection. Two clean approaches: (A) use `store.dispatch(oidcApi.util.updateSubscriptionOptions(...))` + manually add reducer via `store.injectReducers` if RTK exposes it, or (B) design `createClientStore` / `createJourneyStore` to pre-register all possible reducers with their slices (oidc's included) so the store is fully wired before oidc() receives it. Option B is simpler but slightly over-provisions; Option A is lazy but requires RTK 2.0 `combineSlices`.",
-      "answerMode": "single-custom",
-      "recommendedAnswer": "Pre-register all reducers in the owning store (davinci/journey) — avoids runtime injection complexity, reducer keys don't collide (config/node/davinci/journeyReducer/oidc/wellknown are all unique), and the extra state is negligible.",
-      "recommendedOptionIds": ["pre-register"],
-      "options": [
-        {
-          "id": "pre-register",
-          "label": "Pre-register oidc reducers in davinci/journey store upfront"
-        },
-        {
-          "id": "lazy-inject",
-          "label": "Lazy inject via RTK combineSlices at oidc() init time"
-        }
-      ],
-      "required": true
-    },
-    {
-      "id": "scope",
-      "prompt": "Which packages are in scope for this change?",
-      "description": "davinci-client and journey-client are the store owners; oidc-client is the joiner. protect and device-client don't use Redux stores currently.",
-      "answerMode": "multi-custom",
-      "recommendedAnswer": "davinci-client, journey-client, oidc-client only.",
-      "recommendedOptionIds": ["davinci", "journey", "oidc"],
-      "options": [
-        { "id": "davinci", "label": "davinci-client" },
-        { "id": "journey", "label": "journey-client" },
-        { "id": "oidc", "label": "oidc-client" },
-        { "id": "sdk-effects", "label": "sdk-effects (wellknown extraction)" }
-      ],
-      "required": true
-    },
-    {
-      "id": "breaking-changes",
-      "prompt": "Are any breaking changes acceptable?",
-      "description": "Exposing `store` on the return object is purely additive. Adding `store?` to oidc's options bag is additive. The only potential breaking change is if `wellknownApi` is extracted to sdk-effects and packages previously imported it directly — but that's an internal import path change, not a public API change.",
-      "answerMode": "single",
-      "recommendedAnswer": "No breaking changes to public API — internal import path changes are acceptable.",
-      "recommendedOptionIds": ["no-breaking"],
-      "options": [
-        { "id": "no-breaking", "label": "No breaking public API changes" },
-        {
-          "id": "minor-ok",
-          "label": "Minor breaking changes acceptable (semver minor)"
-        }
-      ],
-      "required": true
-    }
-  ]
-}
