+++ packages/davinci-client/src/lib/node.reducer.ts
@@ -37,7 +37,7 @@ import {
   returnQrCodeCollector,
   returnImageCollector,
 } from './collector.utils.js';
-import { resolveCollectorUpdateValue } from './client.store.utils.js';
+import { resolveCollectorUpdateValue } from './collector.resolver.js';
 import type { DaVinciField, UnknownField } from './davinci.types.js';
 import type { PhoneNumberOutputValue, PhoneNumberExtensionOutputValue } from './collector.types.js';
 import type {
